backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "quangtran"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (==0.104.1)",
    "uvicorn (==0.23.2)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "sqlalchemy (==2.0.23)",
    "python-jose (==3.3.0)",
    "passlib (==1.7.4)",
    "python-multipart (==0.0.6)",
    "python-dotenv (==1.0.0)",
    "emails (==0.6)",
    "jinja2 (==3.1.2)",
    "pymysql (==1.1.0)",
    "alembic (==1.12.1)",
    "pydantic-settings (>=2.8.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
